[level 2] Title: 멀리 뛰기, Time: 0.11 ms, Memory: 79 MB -BaekjoonHub

+++ 프로그래머스/lv2/12914. 멀리 뛰기/멀리 뛰기.java
@@ -3,7 +3,7 @@ import java.util.*;
 class Solution {
     
     public long solution(int n) {
-        long[] result = new long[n];
+        int[] result = new int[n];
         
         for(int i = 0; i<n; i++){
             if(i==0) result[i] = 1;
